+++ dolomite_engine/hf_models/__init__.py
@@ -1,4 +1,5 @@
 from .enums import AttentionHeadType, PositionEmbeddingType
+from .model_conversion import export_to_huggingface, import_from_huggingface
 from .modeling_utils_TP import (
     CUDA_RNGStatesTracker,
     get_tensor_parallel_group_manager,
@@ -21,12 +22,6 @@ from .models import (
     MoEDolomiteForCausalLM,
     MoEDolomiteModel,
     convert_gpt_dolomite_to_gpt_crosslayer,
-    export_to_huggingface_bigcode,
-    export_to_huggingface_llama,
-    export_to_huggingface_mixtral,
-    import_from_huggingface_bigcode,
-    import_from_huggingface_llama,
-    import_from_huggingface_mixtral,
 )
 from .register_hf import is_custom_model, register_model_classes
