+++ dolomite_engine/hf_models/model_conversion/llama.py
@@ -1,4 +1,4 @@
-from transformers import AutoConfig, AutoTokenizer, GenerationConfig, MixtralConfig
+from transformers import AutoConfig, AutoTokenizer, GenerationConfig, LlamaConfig
 
 from ...utils import SafeTensorsWeightsManager
 from ..enums import AttentionHeadType
@@ -6,12 +6,12 @@ from ..modeling_utils import (
     interleave_query_key_value_tensor_for_attention,
     split_query_key_value_tensor_for_attention,
 )
+from ..models import GPTDolomiteConfig
+from ..models.gpt_dolomite import interleave_up_gate_tensor_for_mlp, split_up_gate_tensor_for_mlp
 from ..utils import download_repo
-from .gpt_dolomite import interleave_up_gate_tensor_for_mlp, split_up_gate_tensor_for_mlp
-from .moe_dolomite import MoEDolomiteConfig
 
 
-def import_from_huggingface_mixtral(pretrained_model_name_or_path: str, save_path: str) -> None:
+def import_from_huggingface_llama(pretrained_model_name_or_path: str, save_path: str) -> None:
     original_config, tokenizer, downloaded_model_path = download_repo(pretrained_model_name_or_path)
     config = _import_config_from_huggingface(original_config)
 
@@ -19,7 +19,6 @@ def import_from_huggingface_mixtral(pretrained_model_name_or_path: str, save_pat
     state_dict = _import_state_dict_from_huggingface(
         safetensors_weight_manager,
         config.n_layer,
-        config.num_experts,
         config.n_head,
         config.num_key_value_heads,
         config.n_embd // config.n_head,
@@ -36,7 +35,7 @@ def import_from_huggingface_mixtral(pretrained_model_name_or_path: str, save_pat
         tokenizer.save_pretrained(save_path, legacy_format=False)
 
 
-def _import_config_from_huggingface(original_config: MixtralConfig) -> MoEDolomiteConfig:
+def _import_config_from_huggingface(original_config: LlamaConfig) -> GPTDolomiteConfig:
     assert original_config.hidden_act == "silu"
 
     if original_config.num_attention_heads == original_config.num_key_value_heads:
@@ -46,7 +45,9 @@ def _import_config_from_huggingface(original_config: MixtralConfig) -> MoEDolomi
     elif original_config.num_attention_heads > original_config.num_key_value_heads:
         attention_head_type = "gqa"
 
-    config = MoEDolomiteConfig(
+    assert original_config.mlp_bias == original_config.attention_bias
+
+    config = GPTDolomiteConfig(
         vocab_size=original_config.vocab_size,
         n_positions=original_config.max_position_embeddings,
         n_embd=original_config.hidden_size,
@@ -60,15 +61,12 @@ def _import_config_from_huggingface(original_config: MixtralConfig) -> MoEDolomi
         normalization_function="rmsnorm",
         layer_norm_epsilon=original_config.rms_norm_eps,
         use_cache=original_config.use_cache,
-        add_bias=False,
+        add_bias=original_config.attention_bias,
         tie_word_embeddings=original_config.tie_word_embeddings,
         initializer_range=original_config.initializer_range,
         rope_theta=original_config.rope_theta,
+        rope_scaling=original_config.rope_scaling,
         attn_pdrop=original_config.attention_dropout,
-        num_experts=original_config.num_local_experts,
-        num_experts_per_tok=original_config.num_experts_per_tok,
-        output_router_logits=original_config.output_router_logits,
-        router_aux_loss_coef=original_config.router_aux_loss_coef,
         bos_token_id=original_config.bos_token_id,
         eos_token_id=original_config.eos_token_id,
         pad_token_id=original_config.pad_token_id,
@@ -80,7 +78,6 @@ def _import_config_from_huggingface(original_config: MixtralConfig) -> MoEDolomi
 def _import_state_dict_from_huggingface(
     safetensors_weight_manager: SafeTensorsWeightsManager,
     num_layers: int,
-    num_experts: int,
     num_heads: int,
     num_key_value_heads: int,
     head_dim: int,
@@ -102,25 +99,22 @@ def _import_state_dict_from_huggingface(
             f"model.layers.{layer_idx}.post_attention_layernorm.weight"
         )
 
-        state_dict[f"transformer.h.{layer_idx}.mlp.gate.weight"] = safetensors_weight_manager.get_tensor(
-            f"model.layers.{layer_idx}.block_sparse_moe.gate.weight"
+        state_dict[f"transformer.h.{layer_idx}.mlp.c_fc.weight"] = interleave_up_gate_tensor_for_mlp(
+            safetensors_weight_manager.get_tensor(f"model.layers.{layer_idx}.mlp.up_proj.weight"),
+            safetensors_weight_manager.get_tensor(f"model.layers.{layer_idx}.mlp.gate_proj.weight"),
         )
-
-        for expert_idx in range(num_experts):
-            state_dict[f"transformer.h.{layer_idx}.mlp.experts.{expert_idx}.c_fc.weight"] = (
-                interleave_up_gate_tensor_for_mlp(
-                    safetensors_weight_manager.get_tensor(
-                        f"model.layers.{layer_idx}.block_sparse_moe.experts.{expert_idx}.w3.weight"
-                    ),
-                    safetensors_weight_manager.get_tensor(
-                        f"model.layers.{layer_idx}.block_sparse_moe.experts.{expert_idx}.w1.weight"
-                    ),
-                )
+        if f"model.layers.{layer_idx}.mlp.up_proj.bias" in safetensors_weight_manager:
+            state_dict[f"transformer.h.{layer_idx}.mlp.c_fc.bias"] = interleave_up_gate_tensor_for_mlp(
+                safetensors_weight_manager.get_tensor(f"model.layers.{layer_idx}.mlp.up_proj.bias"),
+                safetensors_weight_manager.get_tensor(f"model.layers.{layer_idx}.mlp.gate_proj.bias"),
             )
-            state_dict[f"transformer.h.{layer_idx}.mlp.experts.{expert_idx}.c_proj.weight"] = (
-                safetensors_weight_manager.get_tensor(
-                    f"model.layers.{layer_idx}.block_sparse_moe.experts.{expert_idx}.w2.weight"
-                )
+
+        state_dict[f"transformer.h.{layer_idx}.mlp.c_proj.weight"] = safetensors_weight_manager.get_tensor(
+            f"model.layers.{layer_idx}.mlp.down_proj.weight"
+        )
+        if f"model.layers.{layer_idx}.mlp.down_proj.bias" in safetensors_weight_manager:
+            state_dict[f"transformer.h.{layer_idx}.mlp.c_proj.bias"] = safetensors_weight_manager.get_tensor(
+                f"model.layers.{layer_idx}.mlp.down_proj.bias"
             )
 
         state_dict[f"transformer.h.{layer_idx}.attn.c_attn.weight"] = interleave_query_key_value_tensor_for_attention(
@@ -132,22 +126,38 @@ def _import_state_dict_from_huggingface(
             head_dim,
             attention_head_type,
         )
+        if f"model.layers.{layer_idx}.self_attn.q_proj.bias" in safetensors_weight_manager:
+            state_dict[f"transformer.h.{layer_idx}.attn.c_attn.bias"] = (
+                interleave_query_key_value_tensor_for_attention(
+                    safetensors_weight_manager.get_slice(f"model.layers.{layer_idx}.self_attn.q_proj.bias"),
+                    safetensors_weight_manager.get_slice(f"model.layers.{layer_idx}.self_attn.k_proj.bias"),
+                    safetensors_weight_manager.get_slice(f"model.layers.{layer_idx}.self_attn.v_proj.bias"),
+                    num_heads,
+                    num_key_value_heads,
+                    head_dim,
+                    attention_head_type,
+                )
+            )
+
         state_dict[f"transformer.h.{layer_idx}.attn.c_proj.weight"] = safetensors_weight_manager.get_tensor(
             f"model.layers.{layer_idx}.self_attn.o_proj.weight"
         )
+        if f"model.layers.{layer_idx}.self_attn.o_proj.bias" in safetensors_weight_manager:
+            state_dict[f"transformer.h.{layer_idx}.attn.c_proj.bias"] = safetensors_weight_manager.get_tensor(
+                f"model.layers.{layer_idx}.self_attn.o_proj.bias"
+            )
 
     return state_dict
 
 
-def export_to_huggingface_mixtral(pretrained_model_name_or_path: str, save_path: str) -> None:
-    config: MoEDolomiteConfig = AutoConfig.from_pretrained(pretrained_model_name_or_path)
+def export_to_huggingface_llama(pretrained_model_name_or_path: str, save_path: str) -> None:
+    config: GPTDolomiteConfig = AutoConfig.from_pretrained(pretrained_model_name_or_path)
     original_config = _export_config_to_huggingface(config)
 
     safetensors_weight_manager = SafeTensorsWeightsManager(pretrained_model_name_or_path)
     state_dict = _export_state_dict_to_huggingface(
         safetensors_weight_manager,
         config.n_layer,
-        config.num_experts,
         config.n_head,
         config.num_key_value_heads,
         config.n_embd // config.n_head,
@@ -167,13 +177,12 @@ def export_to_huggingface_mixtral(pretrained_model_name_or_path: str, save_path:
         pass
 
 
-def _export_config_to_huggingface(config: MoEDolomiteConfig) -> MixtralConfig:
+def _export_config_to_huggingface(config: GPTDolomiteConfig) -> LlamaConfig:
     assert config.activation_function == "swiglu"
     assert config.normalization_function == "rmsnorm"
     assert config.position_embedding_type == "rope"
-    assert config.normalize_expert_weights
 
-    original_config = MixtralConfig(
+    original_config = LlamaConfig(
         vocab_size=config.vocab_size,
         max_position_embeddings=config.n_positions,
         hidden_size=config.n_embd,
@@ -184,15 +193,13 @@ def _export_config_to_huggingface(config: MoEDolomiteConfig) -> MixtralConfig:
         hidden_act="silu",
         rms_norm_eps=config.layer_norm_epsilon,
         use_cache=config.use_cache,
+        attention_bias=config.add_bias,
         tie_word_embeddings=config.tie_word_embeddings,
         initializer_range=config.initializer_range,
         rope_theta=config.rope_theta,
         rope_scaling=config.rope_scaling,
         attention_dropout=config.attn_pdrop,
-        num_local_experts=config.num_experts,
-        num_experts_per_tok=config.num_experts_per_tok,
-        output_router_logits=config.output_router_logits,
-        router_aux_loss_coef=config.router_aux_loss_coef,
+        mlp_bias=config.add_bias,
         bos_token_id=config.bos_token_id,
         eos_token_id=config.eos_token_id,
         pad_token_id=config.pad_token_id,
@@ -204,7 +211,6 @@ def _export_config_to_huggingface(config: MoEDolomiteConfig) -> MixtralConfig:
 def _export_state_dict_to_huggingface(
     safetensors_weight_manager: SafeTensorsWeightsManager,
     num_layers: int,
-    num_experts: int,
     num_heads: int,
     num_key_value_heads: int,
     head_dim: int,
@@ -226,23 +232,25 @@ def _export_state_dict_to_huggingface(
             safetensors_weight_manager.get_tensor(f"transformer.h.{layer_idx}.ln_2.weight")
         )
 
-        state_dict[f"model.layers.{layer_idx}.block_sparse_moe.gate.weight"] = safetensors_weight_manager.get_tensor(
-            f"transformer.h.{layer_idx}.mlp.gate.weight"
+        up_weight, gate_weight = split_up_gate_tensor_for_mlp(
+            safetensors_weight_manager.get_tensor(f"transformer.h.{layer_idx}.mlp.c_fc.weight")
         )
+        state_dict[f"model.layers.{layer_idx}.mlp.up_proj.weight"] = up_weight
+        state_dict[f"model.layers.{layer_idx}.mlp.gate_proj.weight"] = gate_weight
 
-        for expert_idx in range(num_experts):
-            up_weight, gate_weight = split_up_gate_tensor_for_mlp(
-                safetensors_weight_manager.get_tensor(
-                    f"transformer.h.{layer_idx}.mlp.experts.{expert_idx}.c_fc.weight"
-                )
+        if f"transformer.h.{layer_idx}.mlp.c_fc.bias" in safetensors_weight_manager:
+            up_bias, gate_bias = split_up_gate_tensor_for_mlp(
+                safetensors_weight_manager.get_tensor(f"transformer.h.{layer_idx}.mlp.c_fc.bias")
             )
-            state_dict[f"model.layers.{layer_idx}.block_sparse_moe.experts.{expert_idx}.w3.weight"] = up_weight
-            state_dict[f"model.layers.{layer_idx}.block_sparse_moe.experts.{expert_idx}.w1.weight"] = gate_weight
+            state_dict[f"model.layers.{layer_idx}.mlp.up_proj.bias"] = up_bias
+            state_dict[f"model.layers.{layer_idx}.mlp.gate_proj.bias"] = gate_bias
 
-            state_dict[f"model.layers.{layer_idx}.block_sparse_moe.experts.{expert_idx}.w2.weight"] = (
-                safetensors_weight_manager.get_tensor(
-                    f"transformer.h.{layer_idx}.mlp.experts.{expert_idx}.c_proj.weight"
-                )
+        state_dict[f"model.layers.{layer_idx}.mlp.down_proj.weight"] = safetensors_weight_manager.get_tensor(
+            f"transformer.h.{layer_idx}.mlp.c_proj.weight"
+        )
+        if f"transformer.h.{layer_idx}.mlp.c_proj.bias" in safetensors_weight_manager:
+            state_dict[f"model.layers.{layer_idx}.mlp.down_proj.bias"] = safetensors_weight_manager.get_tensor(
+                f"transformer.h.{layer_idx}.mlp.c_proj.bias"
             )
 
         query_weight, key_weight, value_weight = split_query_key_value_tensor_for_attention(
@@ -256,8 +264,24 @@ def _export_state_dict_to_huggingface(
         state_dict[f"model.layers.{layer_idx}.self_attn.k_proj.weight"] = key_weight
         state_dict[f"model.layers.{layer_idx}.self_attn.v_proj.weight"] = value_weight
 
+        if f"transformer.h.{layer_idx}.attn.c_attn.bias" in safetensors_weight_manager:
+            query_bias, key_bias, value_bias = split_query_key_value_tensor_for_attention(
+                safetensors_weight_manager.get_tensor(f"transformer.h.{layer_idx}.attn.c_attn.bias"),
+                num_heads,
+                num_key_value_heads,
+                head_dim,
+                attention_head_type,
+            )
+            state_dict[f"model.layers.{layer_idx}.self_attn.q_proj.bias"] = query_bias
+            state_dict[f"model.layers.{layer_idx}.self_attn.k_proj.bias"] = key_bias
+            state_dict[f"model.layers.{layer_idx}.self_attn.v_proj.bias"] = value_bias
+
         state_dict[f"model.layers.{layer_idx}.self_attn.o_proj.weight"] = safetensors_weight_manager.get_tensor(
             f"transformer.h.{layer_idx}.attn.c_proj.weight"
         )
+        if f"transformer.h.{layer_idx}.attn.c_proj.bias" in safetensors_weight_manager:
+            state_dict[f"model.layers.{layer_idx}.self_attn.o_proj.bias"] = safetensors_weight_manager.get_tensor(
+                f"transformer.h.{layer_idx}.attn.c_proj.bias"
+            )
 
     return state_dict
