+++ tools/export_to_hf.py
@@ -1,8 +1,8 @@
-from dolomite_engine.hf_models import export_to_huggingface_llama
+from dolomite_engine.hf_models import export_to_huggingface
 
 
 load_path = "load/"
 save_path = "save/"
 
 # export to HF llama
-export_to_huggingface_llama(load_path, save_path)
+export_to_huggingface(load_path, save_path, model_type="llama")
